requirements.txt
asgiref==3.7.2
azure-core==1.28.0
azure-storage-blob==12.17.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
defusedxml==0.7.1
Deprecated==1.2.14
Django==4.1.7
django-allauth==0.54.0
django-auditlog==2.3.0
django-cors-headers==4.2.0
django-filter==23.2
django-oauth-toolkit==2.3.0
django-simple-history==3.3.0
django-storages==1.13.2
django-tenants==3.5.0
djangorestframework==3.14.0
drf-yasg==1.21.7
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
isodate==0.6.1
Jinja2==3.1.2
jwcrypto==1.5.0
Markdown==3.4.4
MarkupSafe==2.1.3
oauthlib==3.2.2
packaging==23.1
Pillow==10.0.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.4
whitenoise==6.5.0
wrapt==1.15.0
